app/requirements.txt
Flask==2.3.2
pymongo==4.6.0
psycopg2-binary==2.9.3
werkzeug>=2.3.3
SQLAlchemy==1.4.32
flask-sqlalchemy==2.5.1
flask-bcrypt==1.0.1
flask-cors==3.0.10